Create README.md
This repository includes the simple test for DemoQA page. We will fill the form and then submit it. Then we will check, if our input information was added successfully. There is two tests for Chrome and Firefox browsers written with mocha and the final report for the tests. You can run the tests by npm test command in terminal.